requirements.txt
requests==2.31.0
numpy==1.24.3
pandas==2.0.2
SQLAlchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz